refactor: rename /images to /assets
