+++ src/assets/icons/telegram-icon.png
