+++ src/assets/icons/whitepaper-icon.png
